ByteTrack/tools/trt.py
  Add tag support

ByteTrack/train_corners_players.sh
  Add new task

ByteTrack/yolox/core/trainer.py
  Multi-class eval

ByteTrack/yolox/data/datasets/fullcourt.py
  error checking

ByteTrack/yolox/evaluators/__init__.py
ByteTrack/yolox/evaluators/coco_evaluator.py
  Support working coco evaluator

ByteTrack/yolox/tracker/byte_tracker.py
ByteTrack/yolox/tracker/matching.py
  numpy fixes

+++ yolox/tracker/byte_tracker.py
@@ -15,7 +15,7 @@ class STrack(BaseTrack):
     def __init__(self, tlwh, score):
 
         # wait activate
-        self._tlwh = np.asarray(tlwh, dtype=np.float)
+        self._tlwh = np.asarray(tlwh, dtype=np.float32)
         self.kalman_filter = None
         self.mean, self.covariance = None, None
         self.is_activated = False
@@ -106,7 +106,7 @@ class STrack(BaseTrack):
         """Get delta x and delta y
         """
         if self.mean is None:
-            return np.asarray([0, 0], dtype=np.float)
+            return np.asarray([0, 0], dtype=np.float32)
         return self.mean[4:6].copy()
 
     @property
